testing: Add StepAssertSet and StepAPICall
This step allows the assertion that particular structure members should
have a non-zero value, but does not assert against the actual value.
This is useful for allowing tests of data which cannot be set up by the
testing framework to be portable across cloud accounts (for example, the
GetAccount call).

We also allow API calls to be implemented as simple call-back funcs to
avoid the proliferation of test steps - and port all the existing tests
to use that. All are passing.