breaking: Introduce client-wide context.Context
This commit makes every method to every Triton client take a context in
order that the consumer can set deadlines, timeouts and so forth. This
also mandates the removal of the RetryableHTTP library in favour of the
standard library HTTP client (without upstreaming changes to RetryableHTTP).

This is a breaking change, however it is a vast improvement and worth
taking the hit early on.