Demote readers from an `io.ReadSeaker` to an `io.Reader`.
When `triton-go` used a retriable HTTP client, it was required to
use an `io.ReadSeaker`.  Now that `triton-go` doesn't use retriablehttp,
relax this requirement to be just an `io.Reader`.  In the future we
may revisit this and re-introduce a retriable client and promote this
to an `io.ReadSeaker`, or figure out another way of passing in
a Seaker.  The reason for wanting to do this is now you can read from
a pipe (i.e. stdin).