storage: Introduce Objects.GetInfo for sending HEAD requests at Manta (#75)
This introduces two functions under the storage package, Objects.GetInfo and a
convenience wrapper called Objects.IsDir. Objects.GetInfo performs the same
action as Objects.Get except that it uses the http.MethodHead type request. All
headers are returned akin to GetObjectOutput except for the response
body. Objects.IsDir returns either a boolean, whether or not the object is a
directory, and an error in case the Objects.GetInfo call failed.

Fixes #74