Add support for the SetRoleTags call available in the CloudAPI.
This commit adds support for the SetRoleTags call available in the
CloudAPI, which allows for setting a single or multiple role tags
to use with the RBAC, either for an individual resource or for the
whole group of resources (e.g., all the current instances, etc.).

Additionally, this commit also adds GetRoleTags function, whilst not
provided by the CloudAPI, can be used to retrieve a list of currently
set role tags.

Signed-off-by: Krzysztof Wilczynski <kw@linux.com>