refactor: In preparation for triton-cli move triton to subpackage
There was a triton package in the root of the application. In order to
prepare for the creation of a cli, we need main package to be in the
root of the application.

Therefore, the 2 triton package files were moved to the triton subfolder
to form it's own package. Then the associated paths in the other files
were updated